ad
LC Medium
needed help on this one, and I did not understand the answers to I left it alone. Hours later I had the epiphany as to what it all meant, and so I wrote my solution from that understanding and added comments to help me put those thoughts into practice. Yay learning!